+++ src/Application/Command/VideoConference/RequestAccessHandler.php
@@ -14,12 +14,18 @@ use Proximum\Vimeet\Application\Adapter\VideoConferenceAdapterInterface;
 use Proximum\Vimeet\Application\Components\Visio\VisioSettingsRetriever;
 use Proximum\Vimeet\Application\Exception\VideoConference\InvalidTokenGeneratorArgumentsException;
 use Proximum\Vimeet\Application\View\Meeting\VideoConferenceView;
+use Proximum\Vimeet\Domain\Model\Meeting;
+use Proximum\Vimeet\Domain\Model\Participant;
 use Proximum\Vimeet\Domain\Model\VideoConference;
 use Proximum\Vimeet\Domain\Model\VideoConferenceToken;
+use Proximum\Vimeet\Domain\Repository\MeetingRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\VideoConferenceRepositoryInterface;
 
 class RequestAccessHandler
 {
+    /** @var MeetingRepositoryInterface */
+    private $meetingRepository;
+
     /** @var VideoConferenceRepositoryInterface */
     private $videoConferenceRepository;
 
@@ -30,10 +36,12 @@ class RequestAccessHandler
     private $visioSettingsRetriever;
 
     public function __construct(
+        MeetingRepositoryInterface $meetingRepository,
         VideoConferenceAdapterInterface $videoConferenceAdapter,
         VideoConferenceRepositoryInterface $videoConferenceRepository,
         VisioSettingsRetriever $visioSettingsRetriever
     ) {
+        $this->meetingRepository = $meetingRepository;
         $this->videoConferenceRepository = $videoConferenceRepository;
         $this->videoConferenceAdapter = $videoConferenceAdapter;
         $this->visioSettingsRetriever = $visioSettingsRetriever;
@@ -48,6 +56,8 @@ class RequestAccessHandler
      */
     public function handle(RequestAccess $requestAccess): VideoConferenceView
     {
+        $this->addParticipantToMeeting($requestAccess->participant, $requestAccess->meeting);
+
         $videoConference = $this->videoConferenceRepository->findByMeeting($requestAccess->meeting);
         $visioSettings = $this->visioSettingsRetriever->get($requestAccess->meeting->getEvent());
 
@@ -108,5 +118,15 @@ class RequestAccessHandler
             $visioSettings->getEndImage($requestAccess->locale),
             $visioSettings->getEndMessage($requestAccess->locale)
         );
+    }
+
+    private function addParticipantToMeeting(Participant $participant, Meeting $meeting): void
+    {
+        if ($meeting->hasParticipant($participant)) {
+            return;
+        }
+
+        $meeting->addParticipant($participant);
+        $this->meetingRepository->set($meeting);
     }
 }