+++ src/Domain/Model/Meeting.php
@@ -474,6 +474,27 @@ class Meeting implements MessageSubjectInterface
         return [];
     }
 
+    public function addParticipant(Participant $participant): void
+    {
+        if ($participant->getSheet()->getId() === $this->fromSheet->getId()) {
+            $this->fromParticipants->add($participant);
+
+            return;
+        }
+
+        if ($participant->getSheet()->getId() === $this->toSheet->getId()) {
+            $this->toParticipants->add($participant);
+
+            return;
+        }
+
+        throw new \InvalidArgumentException('Participant sheet not known for this meeting');
+    }
+
+    public function hasParticipant(Participant $participant): bool
+    {
+       return in_array($participant, $this->getAllParticipants(), true);
+    }
 
     public function getCreatedType(): string
     {