+++ tests/Application/Command/VideoConference/RequestAccessHandlerTest.php
@@ -1,12 +1,6 @@
 <?php
 
-/*
- * This file is part of the Proximum Vimeet project.
- *
- * Copyright (C) Proximum
- *
- * @author Elao <contact@elao.com>
- */
+namespace Proximum\Vimeet\Tests\Application\Command\VideoConference;
 
 use OpenTok\Session;
 use PHPUnit\Framework\TestCase;
@@ -18,10 +12,13 @@ use Proximum\Vimeet\Application\View\Meeting\VideoConferenceView;
 use Proximum\Vimeet\Domain\Model\Event;
 use Proximum\Vimeet\Domain\Model\Meeting;
 use Proximum\Vimeet\Domain\Model\MeetingSlot;
+use Proximum\Vimeet\Domain\Model\Participant;
+use Proximum\Vimeet\Domain\Model\Sheet;
 use Proximum\Vimeet\Domain\Model\User;
 use Proximum\Vimeet\Domain\Model\VideoConference;
 use Proximum\Vimeet\Domain\Model\VideoConferenceToken;
 use Proximum\Vimeet\Domain\Model\Visio\VisioSettings;
+use Proximum\Vimeet\Domain\Repository\MeetingRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\VideoConferenceRepositoryInterface;
 
 class RequestAccessHandlerTest extends TestCase
@@ -33,8 +30,15 @@ class RequestAccessHandlerTest extends TestCase
         $slot    = $this->prophesize(MeetingSlot::class);
         $user    = $this->prophesize(User::class);
 
+        $sheet = $this->prophesize(Sheet::class);
+        $participant = $this->prophesize(Participant::class);
+        $participant->getSheet()->shouldBeCalled()->willReturn($sheet->reveal());
+        $participant->getUser()->shouldBeCalled()->willReturn($user->reveal());
+
         $slot->getEnd()->shouldBeCalled()->willReturn($slotEnd);
         $meeting->getSlot()->shouldBeCalled()->willReturn($slot->reveal());
+        $meeting->hasParticipant($participant->reveal())->shouldBeCalled()->willReturn(false);
+        $meeting->addParticipant($participant->reveal())->shouldBeCalled();
 
         // Mock
         $session                   = $this->prophesize(Session::class);
@@ -74,7 +78,11 @@ class RequestAccessHandlerTest extends TestCase
 
         $videoConferenceRepository->add($videoConference)->shouldBeCalled();
 
+        $meetingRepository = $this->prophesize(MeetingRepositoryInterface::class);
+        $meetingRepository->set($meeting->reveal())->shouldBeCalled();
+
         $handler = new RequestAccessHandler(
+            $meetingRepository->reveal(),
             $videoConferenceAdapter->reveal(),
             $videoConferenceRepository->reveal(),
             $visioSettingsRetriever->reveal()
@@ -94,7 +102,7 @@ class RequestAccessHandlerTest extends TestCase
         $videoConferenceView = $handler->handle(
             new RequestAccess(
                 $meeting->reveal(),
-                $user->reveal(),
+                $participant->reveal(),
                 'fr'
             )
         );
@@ -109,8 +117,14 @@ class RequestAccessHandlerTest extends TestCase
         $slot    = $this->prophesize(MeetingSlot::class);
         $user    = $this->prophesize(User::class);
 
+        $sheet = $this->prophesize(Sheet::class);
+        $participant = $this->prophesize(Participant::class);
+        $participant->getSheet()->shouldBeCalled()->willReturn($sheet->reveal());
+        $participant->getUser()->shouldBeCalled()->willReturn($user->reveal());
+
         $slot->getEnd()->shouldBeCalled()->willReturn($slotEnd);
         $meeting->getSlot()->shouldBeCalled()->willReturn($slot->reveal());
+        $meeting->hasParticipant($participant->reveal())->shouldBeCalled()->willReturn(true);
 
         // Mock
         $videoConference           = $this->prophesize(VideoConference::class);
@@ -158,7 +172,11 @@ class RequestAccessHandlerTest extends TestCase
 
         $videoConferenceRepository->set($videoConference)->shouldBeCalled();
 
+        $meetingRepository = $this->prophesize(MeetingRepositoryInterface::class);
+        $meetingRepository->set($meeting->reveal())->shouldNotBeCalled();
+
         $handler = new RequestAccessHandler(
+            $meetingRepository->reveal(),
             $videoConferenceInterface->reveal(),
             $videoConferenceRepository->reveal(),
             $visioSettingsRetriever->reveal()
@@ -178,7 +196,7 @@ class RequestAccessHandlerTest extends TestCase
         $videoConferenceView = $handler->handle(
             new RequestAccess(
                 $meeting->reveal(),
-                $user->reveal(),
+                $participant->reveal(),
                 'fr'
             )
         );